moved logic to handler
